formulario montado con sus componentes
